% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwc_110626}
\alias{gwc_110626}
\title{a snapshot of the EBI/NHGRI GWAS catalog}
\format{
An object of class \code{gwaswloc} of length 946062.
}
\usage{
data(gwc_110626)
}
\description{
a snapshot of the EBI/NHGRI GWAS catalog
}
\note{
requires gwascat package to be available so that the
class definition is present
}
\keyword{datasets}
