% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manh.R
\name{make_manh}
\alias{make_manh}
\title{create a manhattan plot given a study accession and a catalog}
\usage{
make_manh(sid, catalog, ...)
}
\arguments{
\item{sid}{string STUDY ACCESSION}

\item{catalog}{instance of gwasWloc from gwascat package}

\item{\dots}{passed to ggmanh manhattan_plot}
}
\description{
create a manhattan plot given a study accession and a catalog
}
\examples{
data("gwc_110626", package="op2workshop")
make_manh("GCST90002322", gwc_110626)
}
