% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monoi}
\alias{monoi}
\title{an ontologyIndex representation of MONDO}
\format{
ontologyIndex
}
\usage{
data(monoi)
}
\description{
an ontologyIndex representation of MONDO
}
\note{
for convenience of visualization of phenotype networks
}
\keyword{datasets}
